Clean up the discord bot commands
I realised now that I can edit webhooks. However, I'd have to use a different setup method like a manual http request, or a command line option.
I think the command line option would overcomplicate things and it's better to keep the discord option as a discord command.
When you add in the effort of changing this it's definitely better to keep it at is.